Struct aws_sdk_imagebuilder::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for imagebuilder.

This client allows ergonomic access to a imagebuilder-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn cancel_image_creation(&self) -> CancelImageCreation<C, M, R>

pub fn create_component(&self) -> CreateComponent<C, M, R>

pub fn create_container_recipe(&self) -> CreateContainerRecipe<C, M, R>

pub fn create_distribution_configuration(
&self
) -> CreateDistributionConfiguration<C, M, R>

pub fn create_image(&self) -> CreateImage<C, M, R>

pub fn create_image_pipeline(&self) -> CreateImagePipeline<C, M, R>

pub fn create_image_recipe(&self) -> CreateImageRecipe<C, M, R>

pub fn create_infrastructure_configuration(
&self
) -> CreateInfrastructureConfiguration<C, M, R>

pub fn delete_component(&self) -> DeleteComponent<C, M, R>

pub fn delete_container_recipe(&self) -> DeleteContainerRecipe<C, M, R>

pub fn delete_distribution_configuration(
&self
) -> DeleteDistributionConfiguration<C, M, R>

pub fn delete_image(&self) -> DeleteImage<C, M, R>

pub fn delete_image_pipeline(&self) -> DeleteImagePipeline<C, M, R>

pub fn delete_image_recipe(&self) -> DeleteImageRecipe<C, M, R>

pub fn delete_infrastructure_configuration(
&self
) -> DeleteInfrastructureConfiguration<C, M, R>

pub fn get_component(&self) -> GetComponent<C, M, R>

pub fn get_component_policy(&self) -> GetComponentPolicy<C, M, R>

pub fn get_container_recipe(&self) -> GetContainerRecipe<C, M, R>

pub fn get_container_recipe_policy(&self) -> GetContainerRecipePolicy<C, M, R>

pub fn get_distribution_configuration(
&self
) -> GetDistributionConfiguration<C, M, R>

pub fn get_image(&self) -> GetImage<C, M, R>

pub fn get_image_pipeline(&self) -> GetImagePipeline<C, M, R>

pub fn get_image_policy(&self) -> GetImagePolicy<C, M, R>

pub fn get_image_recipe(&self) -> GetImageRecipe<C, M, R>

pub fn get_image_recipe_policy(&self) -> GetImageRecipePolicy<C, M, R>

pub fn get_infrastructure_configuration(
&self
) -> GetInfrastructureConfiguration<C, M, R>

pub fn import_component(&self) -> ImportComponent<C, M, R>

pub fn list_component_build_versions(
&self
) -> ListComponentBuildVersions<C, M, R>

pub fn list_components(&self) -> ListComponents<C, M, R>

pub fn list_container_recipes(&self) -> ListContainerRecipes<C, M, R>

pub fn list_distribution_configurations(
&self
) -> ListDistributionConfigurations<C, M, R>

pub fn list_image_build_versions(&self) -> ListImageBuildVersions<C, M, R>

pub fn list_image_packages(&self) -> ListImagePackages<C, M, R>

pub fn list_image_pipeline_images(&self) -> ListImagePipelineImages<C, M, R>

pub fn list_image_pipelines(&self) -> ListImagePipelines<C, M, R>

pub fn list_image_recipes(&self) -> ListImageRecipes<C, M, R>

pub fn list_images(&self) -> ListImages<C, M, R>

pub fn list_infrastructure_configurations(
&self
) -> ListInfrastructureConfigurations<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn put_component_policy(&self) -> PutComponentPolicy<C, M, R>

pub fn put_container_recipe_policy(&self) -> PutContainerRecipePolicy<C, M, R>

pub fn put_image_policy(&self) -> PutImagePolicy<C, M, R>

pub fn put_image_recipe_policy(&self) -> PutImageRecipePolicy<C, M, R>

pub fn start_image_pipeline_execution(
&self
) -> StartImagePipelineExecution<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_distribution_configuration(
&self
) -> UpdateDistributionConfiguration<C, M, R>

pub fn update_image_pipeline(&self) -> UpdateImagePipeline<C, M, R>

pub fn update_infrastructure_configuration(
&self
) -> UpdateInfrastructureConfiguration<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self