Module aws_sdk_imagebuilder::error

Modules

call_rate_limit_exceeded_exception

See CallRateLimitExceededException

client_exception

See ClientException

forbidden_exception

See ForbiddenException

idempotent_parameter_mismatch_exception

See IdempotentParameterMismatchException

invalid_pagination_token_exception

See InvalidPaginationTokenException

invalid_parameter_combination_exception

See InvalidParameterCombinationException

invalid_parameter_exception

See InvalidParameterException

invalid_parameter_value_exception

See InvalidParameterValueException

invalid_request_exception

See InvalidRequestException

invalid_version_number_exception

See InvalidVersionNumberException

resource_already_exists_exception

See ResourceAlreadyExistsException

resource_dependency_exception

See ResourceDependencyException

resource_in_use_exception

See ResourceInUseException

resource_not_found_exception

See ResourceNotFoundException

service_exception

See ServiceException

service_quota_exceeded_exception

See ServiceQuotaExceededException

service_unavailable_exception

See ServiceUnavailableException

Structs

CallRateLimitExceededException

You have exceeded the permitted request rate for the specific operation.

CancelImageCreationError

ClientException

These errors are usually caused by a client action, such as using an action or resource on

CreateComponentError

CreateContainerRecipeError

CreateDistributionConfigurationError

CreateImageError

CreateImagePipelineError

CreateImageRecipeError

CreateInfrastructureConfigurationError

DeleteComponentError

DeleteContainerRecipeError

DeleteDistributionConfigurationError

DeleteImageError

DeleteImagePipelineError

DeleteImageRecipeError

DeleteInfrastructureConfigurationError

ForbiddenException

You are not authorized to perform the requested operation.

GetComponentError

GetComponentPolicyError

GetContainerRecipeError

GetContainerRecipePolicyError

GetDistributionConfigurationError

GetImageError

GetImagePipelineError

GetImagePolicyError

GetImageRecipeError

GetImageRecipePolicyError

GetInfrastructureConfigurationError

IdempotentParameterMismatchException

You have specified a client token for an operation using parameter values that differ from

ImportComponentError

InvalidPaginationTokenException

You have provided an invalid pagination token in your request.

InvalidParameterCombinationException

You have specified two or more mutually exclusive parameters. Review the error message for

InvalidParameterException

The specified parameter is invalid. Review the available parameters for the API

InvalidParameterValueException

The value that you provided for the specified parameter is invalid.

InvalidRequestException

You have made a request for an action that is not supported by the service.

InvalidVersionNumberException

Your version number is out of bounds or does not follow the required syntax.

ListComponentBuildVersionsError

ListComponentsError

ListContainerRecipesError

ListDistributionConfigurationsError

ListImageBuildVersionsError

ListImagePackagesError

ListImagePipelineImagesError

ListImagePipelinesError

ListImageRecipesError

ListImagesError

ListInfrastructureConfigurationsError

ListTagsForResourceError

PutComponentPolicyError

PutContainerRecipePolicyError

PutImagePolicyError

PutImageRecipePolicyError

ResourceAlreadyExistsException

The resource that you are trying to create already exists.

ResourceDependencyException

You have attempted to mutate or delete a resource with a dependency that prohibits this

ResourceInUseException

The resource that you are trying to operate on is currently in use. Review the message

ResourceNotFoundException

At least one of the resources referenced by your request does not exist.

ServiceException

This exception is thrown when the service encounters an unrecoverable exception.

ServiceQuotaExceededException

You have exceeded the number of permitted resources or operations for this service. For service quotas, see EC2 Image Builder endpoints and quotas.

ServiceUnavailableException

The service is unable to process your request at this time.

StartImagePipelineExecutionError

TagResourceError

UntagResourceError

UpdateDistributionConfigurationError

UpdateImagePipelineError

UpdateInfrastructureConfigurationError

Enums

CancelImageCreationErrorKind

CreateComponentErrorKind

CreateContainerRecipeErrorKind

CreateDistributionConfigurationErrorKind

CreateImageErrorKind

CreateImagePipelineErrorKind

CreateImageRecipeErrorKind

CreateInfrastructureConfigurationErrorKind

DeleteComponentErrorKind

DeleteContainerRecipeErrorKind

DeleteDistributionConfigurationErrorKind

DeleteImageErrorKind

DeleteImagePipelineErrorKind

DeleteImageRecipeErrorKind

DeleteInfrastructureConfigurationErrorKind

GetComponentErrorKind

GetComponentPolicyErrorKind

GetContainerRecipeErrorKind

GetContainerRecipePolicyErrorKind

GetDistributionConfigurationErrorKind

GetImageErrorKind

GetImagePipelineErrorKind

GetImagePolicyErrorKind

GetImageRecipeErrorKind

GetImageRecipePolicyErrorKind

GetInfrastructureConfigurationErrorKind

ImportComponentErrorKind

ListComponentBuildVersionsErrorKind

ListComponentsErrorKind

ListContainerRecipesErrorKind

ListDistributionConfigurationsErrorKind

ListImageBuildVersionsErrorKind

ListImagePackagesErrorKind

ListImagePipelineImagesErrorKind

ListImagePipelinesErrorKind

ListImageRecipesErrorKind

ListImagesErrorKind

ListInfrastructureConfigurationsErrorKind

ListTagsForResourceErrorKind

PutComponentPolicyErrorKind

PutContainerRecipePolicyErrorKind

PutImagePolicyErrorKind

PutImageRecipePolicyErrorKind

StartImagePipelineExecutionErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateDistributionConfigurationErrorKind

UpdateImagePipelineErrorKind

UpdateInfrastructureConfigurationErrorKind