Module aws_sdk_imagebuilder::model

Modules

additional_instance_configuration

See AdditionalInstanceConfiguration

ami

See Ami

ami_distribution_configuration

See AmiDistributionConfiguration

component

See Component

component_configuration

See ComponentConfiguration

component_parameter

See ComponentParameter

component_parameter_detail

See ComponentParameterDetail

component_summary

See ComponentSummary

component_version

See ComponentVersion

container

See Container

container_distribution_configuration

See ContainerDistributionConfiguration

container_recipe

See ContainerRecipe

container_recipe_summary

See ContainerRecipeSummary

distribution

See Distribution

distribution_configuration

See DistributionConfiguration

distribution_configuration_summary

See DistributionConfigurationSummary

ebs_instance_block_device_specification

See EbsInstanceBlockDeviceSpecification

filter

See Filter

image

See Image

image_package

See ImagePackage

image_pipeline

See ImagePipeline

image_recipe

See ImageRecipe

image_recipe_summary

See ImageRecipeSummary

image_state

See ImageState

image_summary

See ImageSummary

image_tests_configuration

See ImageTestsConfiguration

image_version

See ImageVersion

infrastructure_configuration

See InfrastructureConfiguration

infrastructure_configuration_summary

See InfrastructureConfigurationSummary

instance_block_device_mapping

See InstanceBlockDeviceMapping

instance_configuration

See InstanceConfiguration

launch_permission_configuration

See LaunchPermissionConfiguration

launch_template_configuration

See LaunchTemplateConfiguration

logging

See Logging

output_resources

See OutputResources

s3_logs

See S3Logs

schedule

See Schedule

systems_manager_agent

See SystemsManagerAgent

target_container_repository

See TargetContainerRepository

Structs

AdditionalInstanceConfiguration

In addition to your infrastruction configuration, these settings provide an extra layer of

Ami

Details of an Amazon EC2 AMI.

AmiDistributionConfiguration

Define and configure the output AMIs of the pipeline.

Component

A detailed view of a component.

ComponentConfiguration

Configuration details of the component.

ComponentParameter

Contains a key/value pair that sets the named component parameter.

ComponentParameterDetail

Defines a parameter that is used to provide configuration details for the component.

ComponentSummary

A high-level summary of a component.

ComponentVersion

The defining characteristics of a specific version of an Amazon Web Services TOE component.

Container

A container encapsulates the runtime environment for an application.

ContainerDistributionConfiguration

Container distribution settings for encryption, licensing, and sharing in a specific Region.

ContainerRecipe

A container recipe.

ContainerRecipeSummary

A summary of a container recipe

Distribution

Defines the settings for a specific Region.

DistributionConfiguration

A distribution configuration.

DistributionConfigurationSummary

A high-level overview of a distribution configuration.

EbsInstanceBlockDeviceSpecification

Amazon EBS-specific block device mapping specifications.

Filter

A filter name and value pair that is used to return a more specific list of results from a

Image

An Image Builder image. You must specify exactly one recipe for the image – either

ImagePackage

Represents a package installed on an Image Builder image.

ImagePipeline

Details of an image pipeline.

ImageRecipe

An image recipe.

ImageRecipeSummary

A summary of an image recipe.

ImageState

Image state shows the image status and the reason for that status.

ImageSummary

An image summary.

ImageTestsConfiguration

Image tests configuration.

ImageVersion

The defining characteristics of a specific version of an Image Builder image.

InfrastructureConfiguration

Details of the infrastructure configuration.

InfrastructureConfigurationSummary

The infrastructure used when building Amazon EC2 AMIs.

InstanceBlockDeviceMapping

Defines block device mappings for the instance used to configure your image.

InstanceConfiguration

Defines a custom source AMI and block device mapping configurations of an instance

LaunchPermissionConfiguration

Describes the configuration for a launch permission. The launch permission modification request is sent to the

LaunchTemplateConfiguration

Identifies an Amazon EC2 launch template to use for a specific account.

Logging

Logging configuration defines where Image Builder uploads your logs.

OutputResources

The resources produced by this image.

S3Logs

Amazon S3 logging configuration.

Schedule

A schedule configures how often and when a pipeline will automatically create a new image.

SystemsManagerAgent

Contains settings for the SSM agent on your build instance.

TargetContainerRepository

The container repository where the output container image is stored.

Enums

ComponentFormat

ComponentType

ContainerRepositoryService

ContainerType

EbsVolumeType

ImageStatus

ImageType

Ownership

NOTE: Ownership::Self has been renamed to ::SelfValue.

PipelineExecutionStartCondition

PipelineStatus

Platform