Module aws_sdk_imagebuilder::operation

Structs

CancelImageCreation

CancelImageCreation cancels the creation of Image. This operation can only be used on

CreateComponent

Creates a new component that can be used to build, validate, test, and assess your

CreateContainerRecipe

Creates a new container recipe. Container recipes define how images are configured, tested, and assessed.

CreateDistributionConfiguration

Creates a new distribution configuration. Distribution configurations define and configure

CreateImage

Creates a new image. This request will create a new image along with all of the

CreateImagePipeline

Creates a new image pipeline. Image pipelines enable you to automate the creation and

CreateImageRecipe

Creates a new image recipe. Image recipes define how images are configured, tested, and

CreateInfrastructureConfiguration

Creates a new infrastructure configuration. An infrastructure configuration defines the

DeleteComponent

Deletes a component build version.

DeleteContainerRecipe

Deletes a container recipe.

DeleteDistributionConfiguration

Deletes a distribution configuration.

DeleteImage

Deletes an image.

DeleteImagePipeline

Deletes an image pipeline.

DeleteImageRecipe

Deletes an image recipe.

DeleteInfrastructureConfiguration

Deletes an infrastructure configuration.

GetComponent

Gets a component object.

GetComponentPolicy

Gets a component policy.

GetContainerRecipe

Retrieves a container recipe.

GetContainerRecipePolicy

Retrieves the policy for a container recipe.

GetDistributionConfiguration

Gets a distribution configuration.

GetImage

Gets an image.

GetImagePipeline

Gets an image pipeline.

GetImagePolicy

Gets an image policy.

GetImageRecipe

Gets an image recipe.

GetImageRecipePolicy

Gets an image recipe policy.

GetInfrastructureConfiguration

Gets an infrastructure configuration.

ImportComponent

Imports a component and transforms its data into a component document.

ListComponentBuildVersions

Returns the list of component build versions for the specified semantic version.

ListComponents

Returns the list of component build versions for the specified semantic version.

ListContainerRecipes

Returns a list of container recipes.

ListDistributionConfigurations

Returns a list of distribution configurations.

ListImageBuildVersions

Returns a list of image build versions.

ListImagePackages

List the Packages that are associated with an Image Build Version, as determined by Amazon EC2 Systems Manager Inventory at build time.

ListImagePipelineImages

Returns a list of images created by the specified pipeline.

ListImagePipelines

Returns a list of image pipelines.

ListImageRecipes

Returns a list of image recipes.

ListImages

Returns the list of images that you have access to.

ListInfrastructureConfigurations

Returns a list of infrastructure configurations.

ListTagsForResource

Returns the list of tags for the specified resource.

PutComponentPolicy

Applies a policy to a component. We recommend that you call the RAM API CreateResourceShare to share resources. If you call the Image Builder API

PutContainerRecipePolicy

Applies a policy to a container image. We recommend that you call the RAM API CreateResourceShare (https://docs.aws.amazon.com/ram/latest/APIReference/API_CreateResourceShare.html) to share resources. If you call the Image Builder API PutContainerImagePolicy, you must also call the RAM API PromoteResourceShareCreatedFromPolicy (https://docs.aws.amazon.com/ram/latest/APIReference/API_PromoteResourceShareCreatedFromPolicy.html) in order for the resource to be visible to all principals with whom the resource is shared.

PutImagePolicy

Applies a policy to an image. We recommend that you call the RAM API CreateResourceShare to share resources. If you call the Image Builder API

PutImageRecipePolicy

Applies a policy to an image recipe. We recommend that you call the RAM API CreateResourceShare to share resources. If you call the Image Builder API

StartImagePipelineExecution

Manually triggers a pipeline to create an image.

TagResource

Adds a tag to a resource.

UntagResource

Removes a tag from a resource.

UpdateDistributionConfiguration

Updates a new distribution configuration. Distribution configurations define and

UpdateImagePipeline

Updates an image pipeline. Image pipelines enable you to automate the creation and

UpdateInfrastructureConfiguration

Updates a new infrastructure configuration. An infrastructure configuration defines the